Rearrange tests for `summarize_hpo_analysis`.
